refactor: move all fastify type patching to types.ts
