+++ src/auth.ts
@@ -10,19 +10,6 @@ let privateKey: Buffer = undefined;
 
 export type AccessToken = { token: string; expiresIn: Date };
 
-declare module 'fastify' {
-  /* eslint-disable @typescript-eslint/no-unused-vars */
-  interface FastifyRequest {
-    userId?: string;
-    premium?: boolean;
-    roles?: Roles[];
-    service?: boolean;
-    accessToken?: AccessToken;
-  }
-
-  /* eslint-enable @typescript-eslint/no-unused-vars */
-}
-
 interface Options {
   secret: string;
 }
